[fix]: refine code
